requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pillow==10.1.0
weasyprint==60.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
email-validator==2.1.0 